Maintain and use a cache of weak tileWindow references
The goal of the TileWindow cache is to provide an easy way
for multiple AccessWindows to share the same TileWindow, when
appropriate, without having to maintain our own system of
tracking references to the TileWindows.  We just exploit the
Javascript engine's garbage collector.

If weak TileWindow references should happen to persist for a while
after our last reference to a TileWindow is released, that's just
a (small) bonus.  Experience suggests they don't persist for very long.

Why stringify the keys?  All the hype promoting Maps mentions that you can
use objects as keys.  So why aren't we using the objects as keys?  What nearly
all the hype about Maps fails to mention is that Map membership tests use exact
equality (===).  So, they are a good way to associate additional properties
with an object without having to modify the object itself.  However, it means
you can't use a newly constructed object to lookup (access) the entry of a
similarly valued object.