Prefetch summary tiles based on (saved) flick levels
Previously, we always prefetched dl1 and the detail tile at the top-left
of the NG-CHM, even if the map had been viewing a different data layer
and had no detail views at the top-left when saved.

The revised system prefetches the thumbnail tiles for each layer and
permanently associates them with the heat map.  It also prefetches the
summary level tiles for the data layer(s) in the flick control.  (If two
are present, it prefetches the summary tiles for the active layer first.)

If the user changes the layer selected in the inactive flick control, the
new system will prefetch the summary level tiles for the new layer.  This
enables a better flick experience if the user then flicks to that layer.

The new prefetching system uses AccessWindows to prefetch the desired
tiles. This not only reduces code use now but it will eventually enable
prefetched tiles to use the same data management subsystem as other tiles.
E.g. it will allow summary tiles not currently associated with either
flick control to be reclaimed using the same methods as other unused tiles.