Maintain references to all available data for tiles in TileWindows
Adds an array of references to the data for the tiles in the TileWindow.
This will eventually become the "real" (and only) way the system caches
tile data.

TileWindows will store references to tile data that becomes available
during their lifetime in this array. It is never explicitly released.
When the TileWindow is garbage collected, so might the data it references
(if no other TileWindow also references it).

Similarly, a reference to a tile's recently received data must be
preserved until a reference to that data can be saved in the tile
arrays of any relevant TileWindows.  To achieve that, we added a
reference to the tile's data in NEWDATA messages.