Fix issue #422 and perhaps issues #404 and #421
No need for this function to change and restore the map data layer.

Since the restoration is asynchronous (waiting for tiles to load) and
timing is very unpredictable, I think that there's a race condition
with all/some of the detail panels being drawn before the map's data
layer was restored.