Create Build Tag w/ GitHub Actions (#442)
This GitHub Action workflow will add a new tag to the repo 
upon pushes to the 'main' branch. This tag will be based on
the version number in CompatibilityManager.js and existing
build tags in the repo.

For example:

If the current version in CompatibilityManager.js is 2.22.1,
the first tag created will be '2.22.1-build-1'.

With the next push to main, if the CompatibilityManager.js version is
unchanged, the tag will be '2.22.1-build-2'.

With the next push to main, if the CompatibilityManager.js version is
bumped to 2.22.2, the tag will be '2.22.2-build-3'.

I.e. the build number always increases by 1.

The .github/workflows/create=build-tag.yml uses an action that depends
on the particular string in CompatibilityManager.js (see 'start_string'
in that workflow file). So added a little warning in the comments to
help alert our future selves to this little fragility.