Change name: 'BuilderMapGen' -> 'GUIBuilderMapGen'
This change was discussed at the NG-CHM meeting yesterday. I also
changed the 'mapgen' names to be 'guibuildermapgen' for symmetry with
other .jar files. Does anyone have build scripts that depend on the
original names?

Made the corresponding changes in the .yml file for the action, and did
a little reordering of steps.