Show warning message prior to large matrix downloads. See issue #502.
If the number of array elements to download exceeds a threshold (currently
one million array elements) show a warning to the user that the download
may kill the browser due to memory exhaustion.

If the user chooses to proceed, a progress bar is displayed.

If the warning dialog is displayed, the user can cancel the download.

For large downloads there is a noticeable delay between when we have
finished all processing (and hide the dialog) and when the browser is
ready to save the file.  The browser can still crash during this time.
I'm not sure if there's anything that we can do about that.