Change dendrogram scale multiplicatively on scroll. See issue #504.
The issue was caused by the initial dendrogram scale being set to a
value smaller than the minimum scale allowed after a dendrogram rescale.
(I think because the minimum dendrogram leaf height is very large.)

So scrolling would try to reduce the scale, but then increase it to the
minimum allowed value.

Reducing the minimum dendrogram scale would only delay the issue to maps
with even longer minimum leaf heights.

I've also noticed issues on other maps where scaling only makes miniscule
changes to the dendrogram height and it's extremely tedious to get the
desired scale.

So, this patch changes the scroll behavior so that it multiplies the
dendrogram scale for each level of zoom, rather than incrementing/
decrementing the scale by a constant each time.