Change workflow from macOS to ubuntu
The build-NGCHMSupportFiles.yml action failed when 2.24.0 was released.
Analysis showed the failure was due to the code trying to run on what
appears to be a minified version of custom.js, and getting to the line
for the cbioportal logo:

   logo: "https://www.cbioportal.org/images/cbioportal_logo.png",

This line includes the search string hard-coded into
CompilerUtilities.java: 'images/', therefore triggering an attempt to
copy a fail 'cbioportal_logo.png', which of course does not exist in
this repo, resulting in failure.

For reasons I **do not understand**, changing the runner from macOS to
ubuntu eliminates the problem (I tested in a personal fork of this
repo).

I also added the workflow_dispatch trigger to allow this action to be
run manually (in addition to when there is a release) so that we don't
have to create a new release in order to build the artifacts.