Get version from action instead of tag
The build-NGCHMSupportFiles action was getting the version number from
the tag. This is because originally the only trigger for this workflow
was a release. However, I added a manual trigger in
b96016e32eb1f2a45f13ccef1fd2bcd79303f9e7. However, when that manual
trigger is used, the build fails because there is no release tag from
which to collect the version number.

This commit uses the same custom action as
create-tag-and-build-NG-CHM-Artifacts.yml to get the version number from
CompatibilityManager.js:
  https://github.com/marohrdanz/extract-version-from-file