Add log4j and debug compile/build targets
Adding basic log4j infrastructure. Also added compile and build targets
to build_shaidyRmapgen.xml to allow for compiling and creating a .jar
file with debugging info turned on.

Since some of that debugging/logging can affect performance, created
these as separate targets. To use them for creating a version of
ShaidyMapGen.jar with logging and debugging info:

   ant -f build_shaidyRmapgen.xml debug_create_jar

The production build has not changed in any way a user would typically
notice: Added log4j jar files to the /lib, so that the code will compile,
but did NOT include the log4j2.xml. This way the default root logger
will be used, which has no debugging info, and level is set to error. So
for the most part the production build is the same.

There is a way to use separate log4j config files for different builds,
but couldn't get that to work quickly, so leaving that for the future.