Use default box size params if not in saved state
This was done in preparation for implementing an initial panel
configuration from the R package.

It appears the various size-related parameters for a detail map are not
completely independent of each other (e.g. dataBoxHeight,
dataViewHeight, and dataPerRow are interrelated). My plan for the R
package is to define a minimal set of independent parameters, and use
this NG-CHM viewer code to calculate the rest.

For reference, these parameters are the ones that appear in the
"panel_configuration" section of mapConfig.json for a detail pane.