Validate gear before auto apply if special coords
In the special coordinates feature request (#546) we have been
automatically clicking the 'APPLY' button for users to save them a
click. However, in situations were there are special coordinates, but no
discrete coordinates available to pre-populate the 'Color By' dropdown,
clicking the 'APPLY' button pops up a system message alerting the user
about the missing requirements. Probably this will be confusing to the
user (and at the very least annoying). So add a validation of the gear
menu selections before calling `applyPanel()`.

This commit includes a little refactoring of `applyPanel()` to
facilitate the check and (hopefully) make the code 3% easier to parse.

Revert "Validate gear before auto apply if special coords"

This reverts commit ebde3536ddf411f4ef509d05e564932fa2f99127.

I think we're going to take this in a different direction. Reverting for
now, and if the other direction pans out we'll rebase away these
commits.

WIP: Checkpoint for black points option

This commit will be rebased away before any merge into main.

This is just a checkpoint commit. This implementation 'works', but
is not good and causes at least one bug: because it adds a dummy
covariate for the all black points, the preference manager code crashes
if the user tries to, for example, change the heatmap colors.

I have ideas for a different implementation, but saving this checkpoint
in the meantime. (I'm worried my other ideas will have issues with
selections).

WIP: "all black" option

This commit will be rebased away before any merge into main.

This is another step towards adding an option to the gear menu
covariates "Color By" dropdown to make the points all black.

This is an improvement over the previous commit because it does not have
the UserPreferenceManager bug (which is present in the previous commit).

This implementation is still haphazard, and will be cleaned up before any
merge into main.

WIP: squash me

WIP: Add hack so no "All Black" in Group dropdown

This commit will be rebased away before any merge into main.

Adding this hack for the moment while I think of something better.

Squashme

WIP: Clean up setAxisCoCoData

squashme