Combine individual custom script version numbers
The individual custom script version identifiers are concatenated,
separated by semicolons, to make the final custom script version.
Appropriate strings are included for scripts that do not set a version
("undefined") or that fail to load ("error") are included to maintain a
direct correspondence with the list of custom scripts.

Also, a console message is output for each custom script including its
URL and version or error status.