Disable link outs that do not work for non-human bio.species
See enhancement request #551 for more details.

This commit disables link outs for COSMIC, Decipher, DepMap, and Ideogram
Viewer if property "bio.species" is set to anything other than
"Homo_sapiens". (List provided by Faiza.)

We still define the Plugin so that it shows up in the list of installed
plugins, but do not define any link outs so that it does not appear in
the list of active plugins.

Note that the same mechanism could be applied (in reverse) for link outs
that only work for a non-human species.