Allow for panel layout from NGCHM R package
This is the ShaidyMapGen.jar component of the NGCHM R package feature
request of allowing users to implement an initial panel layout from the
R package.

In order to produce a .ngchm file with panels initialized, the NGCHM R
package changes in
https://github.com/MD-Anderson-Bioinformatics/NGCHM-R/pull/75 are also
required.

If this version of ShaidyMapGen.jar is used with NGCHM R code that does
NOT include panel information, a note is printed for the user, and the
map builds w/o panel information. NGCHM R package versions <= 1.0.4 do
not have panel information.