Enable top items to be selected from any continuous covariate
Allows for much more flexible top item selection and future capabilities.

Instead of providing a list of top item labels, the user now chooses a
continuous covariate. The items with the biggest value of that covariate
are the top items.

Old maps that specify a list of labels are upgraded by compatibility
manager. It creates a hidden continuous covariate and sets the default
top item selection to that covariate.

Currently, this provides a superset of the previous top items capabilities
with the exception that the user cannot select an arbitrary set of new
top items interactively.  I propose to implement the long-discussed ability
to create and edit new covariates interactively to restore that capability.
I think we should only release this change together with the ability to
create and edit covariates interactively.

Implementation notes:
- Implemented the core functionality in heatMap.getTopItems.
- Added a function for upgrading old maps in CompatibilityManager. Since
  this requires changes to both mapConfig and mapData, I added a new
  function checkAxis that is called when both become available.
- Changed UserPreferenceManager to display a dropdown selector for letting
  the user choose the continuous covariate to use (or none).