Allow user to edit the types and visibility of label components
Replaces the label_type entry in mapData with a labelTypes entry.
Elements of labelTypes are objects with a "type" (name) text member and
a boolean visible member.
The Compatibility Manager replaces label_type with labelTypes. It sets
only the first entry as visible.

Compatibility manager also splits any type names at '.bar.'. This was
never a part of the specification, but for a long time the R package
generated such entries.

The labelTypes changes are incompatible with the current builder. So it
needs to be upgraded.