Add call to startChange on keydown events
Previously, the user could change a value in the top items text box
or the breakpoint inputs, but the Apply/Reset buttons would not
become active until the user clicked outside the input.

This commit adds keydown event handlers and calls startChange
on keydown events in those inputs, causing the Apply/Reset
buttons to become active if they were not already.

This happens even if the user doesn't end up making a change to
the input, which is a change from the previous behavior but
no different than modifying any other input and then undoing
that change.