Initial implementation of command execution and command window.
Center of the implementation is the EXEC module in ExecCommand.js. In
particular, the function execCommand, which takes an array of "args"
and an output interface and executes the command given by arg[0].

There is also a "runCommand" function that takes a string, parses it into
words (args) and then executes the resulting command.

execCommand is exported to the linkouts.  It uses a console.log based
output interface defined in NGCHM_Util.js.

runCommand is used by CommandMode.js which lets the user enter commands
interactively. It uses an output interface that writes into a textarea
in the popup window.  The popup window is created in UI-Manager.js.

Commands are defined in modules appropriate to the command and are made
known to execCommand by registering themselves using UTIL.registerCommand.

Commands are implemented by "routes" inspired by the Node.js Express
interface. Each function in the path takes three parameters:
- req - the command. Initially, it contains the args array, but this is
  consumed and transformed by middlewares in the path.
- res - the command result. It contains the output interface. Commands that
  return a value set the value field of res.
- next - function to pass control to the next middleware in the route (if
  no parameter) or to another route if specified. In either case, next
  returns after it completes, and the caller should then return if nothing
  further should be executed.

The EXEC module (ExecCommand.js) defines several middlewares that define
tasks that could be commonly useful. For instance, the EXEC.reqColor
function is a middleware that requires a color argument, validates the
argument is a valid color, sets req.color to the color, and consumes the
argument.