Record embedded mode and widgetMode explicitly
Made NgChm.API.embedCHM explicitly record the fact that the NG-CHM is
embedded, rather than inferring that from MMGR.embeddedMapName (which
will change/go away is a subsequent commit).

At some point in the distant past, it appears that it was possible to
set widgetMode to "web".  But for a long time it has always been
compiled into the widget with the fixed value "file".

This commit removes this magic variable in favor of a runtime value
that NgChm.API.embed sets explicitly to "file".  This maintains
the previous behavior.

In a subsequent commit, we will add the capability to set the widgetMode
as an option to NgChm.API.embedCHM.