Removed duplicated keydown handlers for preference tabs
Three out of four preferences tab need a keydown handler to initiate a
change if it appears the user is changing an ngchm-upm-input class element.
The fourth tab will probably need one soon also.  (e.g. let the user
edit the map's description.)

This commit creates a handler when the tab is created, for all tabs.