Added support for multiple NG-CHMs to DetailHeatMaps
This commit focuses on support for multiple NG-CHMs in detail map views.

When a new detail view is created, it will be for the currently selected
heat map.

If a different heat map is selected while the detail view is open, it will
stay open on the same heat map, but it will lose primary status if it had
had it.

If there are existing detail views of the newly selected heat map, one of
them will be promoted to be the primary detail view.  If there are no
existing detail views of the new selected heat map, there will be no primary
detail view.

It is therefore now possible for there to be multiple detail views open,
but no primary (because they are all for heatMaps other than the selected
one). This is a change from previous assumptions that there was always a
primary detail view if there were any detail views.