Added heatmap-specific panel decorations
During meeting discussions it was felt that adding heatmap-specific UI
elements to the panel header would make it easier to know which panels
applied to which heat maps.  This commit implements that request.

Remnants of a differently colored "active" pane header were also removed.
These remnants were inactive because the CSS for the active pane header
was commented out.  Also, the determination of which panel was "active"
was minimal and nowhere near complete. I removed these remnants to avoid
potential confusion in the future.