Made further changes in response to CodeRabbit commits
setAxisCovariateOrder was a separate function from setCovariateOrder
(and the setRowClassificationOrder and setColClassificationOrder
functions it replaced) and so there was no actual brokage.  But,
the two functions are very similar in both name and function and
undoubtedly this would have led to confusion and errors in the
future.  I consolidated both functions into a single function
that takes an optional order parameter.

The jsPDF setDrawColor was working in my tests, but I followed
CodeRabbit's suggestion to ensure compatibility.

I made the suggested changes to improve the robustness of
getCovariateOrder.

I left fixing the redundancy in the totals and accessors API surface
for another day. This is a rabbit hole I don't want to go down just now.